requirements.txt
fastapi
uvicorn
python-multipart
jinja2
sqlalchemy
pydantic
pydantic[dotenv]
pydantic[email]
requests
python-jose[cryptography]